requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.4
h11==0.14.0
idna==3.10
numpy==2.1.3
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
